Organize components by domain (investigation, finding, shared)
