+++ src/components/filters/ConceptSchemeFilter.tsx
@@ -1,4 +1,4 @@
-import { Tooltip } from "../Tooltip";
+import { Tooltip } from "../shared/Tooltip";
 import {
   isSelected,
   toggleConcept,
