+++ src/components/finding/CIVisualization.tsx
