+++ src/components/finding/ComparisonRow.css
