+++ src/components/finding/ComparisonRow.tsx
