+++ src/components/finding/EffectEstimateCard.css
