+++ src/components/finding/EffectEstimateCard.tsx
