+++ src/components/finding/FindingCard.css
