+++ src/components/finding/FindingCard.tsx
