+++ src/components/finding/OutcomeDetails.css
