+++ src/components/finding/SampleDetails.css
