+++ src/components/investigation/AbstractSection.css
