+++ src/components/investigation/AbstractSection.tsx
