+++ src/components/investigation/FindingsSection.css
