+++ tests/components/shared/Pagination.test.tsx
@@ -1,6 +1,6 @@
 import { describe, test, expect, vi } from "vitest";
 import { render, screen, fireEvent } from "@testing-library/preact";
-import { Pagination } from "@/components/Pagination";
+import { Pagination } from "@/components/shared/Pagination";
 
 describe("Pagination", () => {
   test("renders numbered buttons for small totals", () => {
